pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "marine-accident-risk"
version = "0.1.0"
description = "Marine accident risk prediction over grid x time using public OPEN APIs"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "requests>=2.31",
    "lightgbm>=4.0",
    "scikit-learn>=1.3",
    "shap>=0.44",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "streamlit>=1.30",
    "pydeck>=0.8",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "tqdm>=4.66",
]

[project.scripts]
marine-risk = "marine_accident_risk.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"marine_accident_risk" = ["**/*.yaml", "**/*.json"]
